requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
cryptography==40.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-nose==1.4.7
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.7.0
filelock==3.12.0
freetype-py==2.3.0
frozenlist==1.3.1
idna==3.3
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.2
mysql-connector-python==8.0.28
nose==1.3.7
oauthlib==3.2.2
packaging==23.1
pbr==5.11.1
phonenumbers==8.13.8
phonenumberslite==8.13.8
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
protobuf==4.22.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycairo==1.23.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
reportlab==4.0.0
requests==2.28.0
requests-oauthlib==1.3.1
rlPyCairo==0.2.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
stevedore==5.0.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
yarl==1.8.1
